rac ruoi

+++ resources/views/site/cart.php
@@ -22,36 +22,38 @@
         </h3>
       </div>
       <form action="/change_cart" method="post">
-        <?php foreach ($cart as $key => $value) : ?>
-          <div class="flex items-center hover:bg-gray-100 -mx-8 px-6 py-5">
-            <div class="flex w-2/5">
-              <!-- product -->
-              <div class="w-20">
-                <img class="h-24" src="/images/<?php echo $value['images'] ?>" alt="" />
+        <?php if (isset($_SESSION['products'])) : ?>
+          <?php foreach ($cart as $key => $value) : ?>
+            <div class="flex items-center hover:bg-gray-100 -mx-8 px-6 py-5">
+              <div class="flex w-2/5">
+                <!-- product -->
+                <div class="w-20">
+                  <img class="h-24" src="/images/<?php echo $value['images'] ?>" alt="" />
+                </div>
+                <div class="flex flex-col justify-between ml-4 flex-grow">
+                  <span class="font-bold text-sm"><?php echo $value['title'] ?></span>
+                  <a href="/delete_cart?id=<?= $key ?>" class="font-semibold hover:text-red-500 text-gray-500 text-xs">Xóa</a>
+                </div>
               </div>
-              <div class="flex flex-col justify-between ml-4 flex-grow">
-                <span class="font-bold text-sm"><?php echo $value['title'] ?></span>
-                <a href="/delete_cart?id=<?= $key ?>" class="font-semibold hover:text-red-500 text-gray-500 text-xs">Xóa</a>
+              <div class="sm:w-40 flex justify-center mx-14">
+                <button type="button" onClick="decreaseCount(event, this)" class="prevNumberProduct w-8 sm:w-1/3 border-2 border-black border-r-0 translate-x-1 rounded-l-[3px]">-</button>
+                <input type="text" class="inputNumberProduct w-8  sm:w-1/3 appearance-none  outline-none border-2 border-black m-0 text-center" value="<?= $value['numberQuantity'] ?>" name="numberQuantity[<?= $key ?>]" id="">
+                <button type="button" class="nextNumberProduct w-8  sm:w-1/3 border-2 border-black border-l-0 -translate-x-1 rounded-r-[3px]" onClick="increaseCount(event, this)">+</button>
               </div>
+              <span class="text-center w-1/5 font-semibold text-sm tracking-wider"><?php echo $value['price'] ?> ₫</span>
+              <span class="text-center w-1/5 font-semibold text-sm tracking-wider"><?php echo  $value['price'] * $value['numberQuantity'] ?> ₫</span>
+              <?php $total +=  $value['price'] * $value['numberQuantity'] ?>
             </div>
-            <div class="sm:w-40 flex justify-center mx-14">
-              <button type="button" onClick="decreaseCount(event, this)" class="prevNumberProduct w-8 sm:w-1/3 border-2 border-black border-r-0 translate-x-1 rounded-l-[3px]">-</button>
-              <input type="text" class="inputNumberProduct w-8  sm:w-1/3 appearance-none  outline-none border-2 border-black m-0 text-center" value="<?= $value['numberQuantity'] ?>" name="numberQuantity[<?= $key ?>]" id="">
-              <button type="button" class="nextNumberProduct w-8  sm:w-1/3 border-2 border-black border-l-0 -translate-x-1 rounded-r-[3px]" onClick="increaseCount(event, this)">+</button>
-            </div>
-            <span class="text-center w-1/5 font-semibold text-sm tracking-wider"><?php echo $value['price'] ?> ₫</span>
-            <span class="text-center w-1/5 font-semibold text-sm tracking-wider"><?php echo  $value['price'] * $value['numberQuantity'] ?> ₫</span>
-            <?php $total +=  $value['price'] * $value['numberQuantity'] ?>
-          </div>
-        <?php endforeach; ?>
-        <button class="flex font-semibold text-indigo-600 text-sm mt-10">Sửa giỏ hàng</button>
+          <?php endforeach; ?>
+          <button class="flex font-semibold text-indigo-600 text-sm mt-10">Sửa giỏ hàng</button>
       </form>
-      <a href="#" class="flex font-semibold text-indigo-600 text-sm mt-10">
-        <svg class="fill-current mr-2 text-indigo-600 w-4" viewBox="0 0 448 512">
-          <path d="M134.059 296H436c6.627 0 12-5.373 12-12v-56c0-6.627-5.373-12-12-12H134.059v-46.059c0-21.382-25.851-32.09-40.971-16.971L7.029 239.029c-9.373 9.373-9.373 24.569 0 33.941l86.059 86.059c15.119 15.119 40.971 4.411 40.971-16.971V296z" />
-        </svg>
-        Tiếp tục mua sắm
-      </a>
+    <?php endif; ?>
+    <a href="#" class="flex font-semibold text-indigo-600 text-sm mt-10">
+      <svg class="fill-current mr-2 text-indigo-600 w-4" viewBox="0 0 448 512">
+        <path d="M134.059 296H436c6.627 0 12-5.373 12-12v-56c0-6.627-5.373-12-12-12H134.059v-46.059c0-21.382-25.851-32.09-40.971-16.971L7.029 239.029c-9.373 9.373-9.373 24.569 0 33.941l86.059 86.059c15.119 15.119 40.971 4.411 40.971-16.971V296z" />
+      </svg>
+      Tiếp tục mua sắm
+    </a>
 
     </div>
